requirements.txt
# Core
pillow>=9.0
numpy>=1.24

# Layout detection (PP-DocLayout-V3)
ultralytics>=8.0
huggingface-hub>=0.20

# OCR model (GLM-OCR)
torch>=2.0
transformers>=4.40

# PDF loading
pymupdf>=1.23

# Optional — speeds up HF downloads
# hf_transfer
